SequenceMatcher better able to handle moved code
This is a big improvement, but I've only just realized
that a part of my approach has been wrongheaded. I need
a way to measure the distance between two nodes, and
can probably discard these ideas about complexity and
depth. A simple levenshtein distance taken on the raw
strings may be enough, or I may have to come up with
some measurement specific to an AST.